Primer commit desde las correciones con Vicko, se agrego la logica de retorno de componentes con map y la ubicacion de cada componente segun el archivo desde dond se renderiza, falta el refactor de ListItems
