+++ src/App.js
@@ -1,87 +1,22 @@
 import React, { useEffect, useState } from "react";
-
+import axios from "axios";
 import "./App.scss";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import Persons from "./components/Persons";
-import axios from "axios";
+import { Card, Button, Row, Col, Container } from "react-bootstrap";
 
-const tasks = [
-  {
-    user: [
-      {
-        id: 1,
-        title: "Título 1",
-        inModification: false,
-        checked: false,
-        name: "task 1",
-        defaulValue: "aaa",
-      },
-      {
-        id: 2,
-        title: "Título 2",
-        inModification: false,
-        checked: true,
-        name: "task 2",
-        defaulValue: "a",
-      },
-      {
-        id: 3,
-        title: "Título 3",
-        inModification: false,
-        checked: false,
-        name: "task 3",
-        defaulValue: "aaaaa",
-      },
-      {
-        id: 4,
-        title: "Título 4",
-        inModification: false,
-        checked: false,
-        name: "task 4",
-        defaulValue: "",
-      },
-      {
-        id: 5,
-        title: "Título 5",
-        inModification: false,
-        checked: true,
-        name: "task 5",
-        defaulValue: "",
-      },
-      {
-        id: 6,
-        title: "Título 6",
-        inModification: false,
-        checked: false,
-        name: "task 6",
-        defaulValue: "",
-      },
-      {
-        id: 7,
-        title: "Título 7",
-        inModification: false,
-        checked: false,
-        name: "task 7",
-        defaulValue: "",
-      },
-    ],
-  },
-];
+import Persons from "./components/Persons";
+import NavbarPage from "./components/NavbarPage";
+import FormNewPerfil from "./components/FormNewPerfil";
 
 const App = () => {
-  const [peopleData, setPeopledata] = useState({
-    data: [""],
-  });
+  const [peopleData, setPeopledata] = useState([]);
   //Pide la informacion a la API se ejecuta una sola vez (se puede mejorar la forma en que se una axios)
   useEffect(() => {
-    //Solo estoy pidiendo la info del user 1
     const fetchData = async () => {
       try {
         const responseGet = await axios.get("https://reqres.in/api/users/");
-        console.log(responseGet.data.data[0], "es del get");
-        let dataGet = responseGet.data.data[0];
-        setPeopledata({ data: [dataGet] });
-        //
+        let users = responseGet.data.data;
+        setPeopledata(users);
       } catch {
         const error = new Error("Valio vrg el GET");
         console.error(error);
@@ -90,14 +25,35 @@ const App = () => {
     fetchData();
   }, []);
 
+  const postData = async (userName, email) => {
+    console.log("la funcion postData");
+    //event.persist();
+    try {
+      const responsePost = await axios.post("https://reqres.in/api/users/", {
+        first_name: userName,
+        email: email,
+      });
+
+      let newUser = responsePost.data;
+      setPeopledata([...peopleData, newUser]);
+    } catch {
+      const error = new Error("Valio vrg el POST");
+      console.error(error);
+    }
+  };
+
   return (
     <Router>
       <Route path="/" exact>
-        <Persons
-          tasks={tasks}
-          peopleData={peopleData}
-          setPeopledata={setPeopledata}
-        />
+        <NavbarPage emailUser="jotolon@gamil.com" />
+        <Row className="align-items-center">
+          {peopleData.map((item, index) => (
+            <Persons key={index} first_name={item.first_name} index={index} />
+          ))}
+        </Row>
+      </Route>
+      <Route path="/newperfil" exact>
+        <FormNewPerfil postData={postData} />
       </Route>
     </Router>
   );
