+++ src/components/Persons.js
@@ -6,29 +6,67 @@ import NavbarPage from "./NavbarPage";
 import FormNewPerfil from "./FormNewPerfil";
 import PersonalTasks from "./PersonalTasks";
 
+const tasks = [
+  {
+    id: 1,
+    title: "Título 1",
+    inModification: false,
+    checked: false,
+    name: "task 1",
+    defaulValue: "aaa",
+  },
+  {
+    id: 2,
+    title: "Título 2",
+    inModification: false,
+    checked: true,
+    name: "task 2",
+    defaulValue: "a",
+  },
+  {
+    id: 3,
+    title: "Título 3",
+    inModification: false,
+    checked: false,
+    name: "task 3",
+    defaulValue: "aaaaa",
+  },
+  {
+    id: 4,
+    title: "Título 4",
+    inModification: false,
+    checked: false,
+    name: "task 4",
+    defaulValue: "",
+  },
+  {
+    id: 5,
+    title: "Título 5",
+    inModification: false,
+    checked: true,
+    name: "task 5",
+    defaulValue: "",
+  },
+  {
+    id: 6,
+    title: "Título 6",
+    inModification: false,
+    checked: false,
+    name: "task 6",
+    defaulValue: "",
+  },
+  {
+    id: 7,
+    title: "Título 7",
+    inModification: false,
+    checked: false,
+    name: "task 7",
+    defaulValue: "",
+  },
+];
+
 const Persons = (props) => {
   //Envia un POST al API (se puede mejorar la forma en que se una axios)
-  const postData = async (userName, email) => {
-    console.log("la funcion postData");
-    //event.persist();
-    try {
-      const responsePost = await axios.post("https://reqres.in/api/users/", {
-        first_name: userName,
-        email: email,
-      });
-      props.tasks.push({
-        userName: [],
-      });
-      console.log(props.tasks);
-      let data = responsePost.data;
-      props.setPeopledata({
-        data: [...props.peopleData.data, data],
-      });
-    } catch {
-      const error = new Error("Valio vrg el POST");
-      console.error(error);
-    }
-  };
 
   const showButton = (numberOfTasks, index) => {
     if (numberOfTasks === 0) {
@@ -45,42 +83,28 @@ const Persons = (props) => {
       );
     }
   };
+  let numberOfTasks = tasks.length;
+  console.log("Es de un solo componente");
 
   return (
     <Router>
       <Route path="/" exact>
-        <NavbarPage emailUser={props.peopleData.data[0].email} />
-        <ul>
-          {props.peopleData.data.map((item, index) => {
-            console.log(props.peopleData.data);
-            let numberOfTasks = props.tasks[0].user.length;
-
-            return (
-              <Container key={index}>
-                <Row>
-                  <Col xs={5} className="mt-2">
-                    <Card key={index} style={{ width: "18rem" }}>
-                      <Card.Header className="text-center">
-                        {item.first_name}
-                      </Card.Header>
-                      <Card.Body>
-                        <Card.Title>
-                          You have {numberOfTasks} tasks to do
-                        </Card.Title>
-                        {showButton(numberOfTasks, index)}
-                      </Card.Body>
-                    </Card>
-                  </Col>
-                </Row>
-              </Container>
-            );
-          })}
-        </ul>
+        <Col className="mx-2 my-2">
+          <Card style={{ width: "18rem" }}>
+            <Card.Header className="text-center">
+              {props.first_name}
+            </Card.Header>
+            <Card.Body>
+              <Card.Title>You have {numberOfTasks} tasks to do</Card.Title>
+              <Link to={`/${props.first_name}/tasks`}>
+                <Button variant="primary">Go Tasks</Button>
+              </Link>
+            </Card.Body>
+          </Card>
+        </Col>
       </Route>
       {/* Estos elementos estan aqui porque necesitan variables de ESTE Componente y del componente App para fucionar */}
-      <Route path="/newperfil" exact>
-        <FormNewPerfil postData={postData} />
-      </Route>
+
       <Route path="/tasks/:id">
         <PersonalTasks tasks={props.tasks} />
       </Route>
